Refactor page layout sharing and tag template
