+++ src/components/Page.js
@@ -1,28 +1,26 @@
 import Blockquote from './Blockquote';
 import CodeColumns from './CodeColumns';
-import DocsetMenu from './DocsetMenu';
 import ExpansionPanel from './ExpansionPanel';
-import Footer from './Footer';
 import InlineCode from './InlineCode';
+import PageLayout, {usePageLayoutProps} from './PageLayout';
 import Pagination from './Pagination';
 import PropTypes from 'prop-types';
-import React, {Fragment, createElement, useCallback, useMemo} from 'react';
-import RelativeLink, {PrimaryLink} from './RelativeLink';
+import React, {Fragment, createElement, useMemo} from 'react';
+import RelativeLink from './RelativeLink';
+import TableOfContents from './TableOfContents';
 import TypeScriptApiBox from './TypeScriptApiBox';
 import VersionBanner from './VersionBanner';
 import autolinkHeadings from 'rehype-autolink-headings';
-import getShareImage from '@jlengstorf/get-share-image';
-import path, {dirname} from 'path';
 import rehypeReact from 'rehype-react';
 import useLocalStorage from 'react-use/lib/useLocalStorage';
 import {
   Box,
   Button,
-  Divider,
-  Flex,
+  HStack,
   Heading,
   ListItem,
   OrderedList,
+  Stack,
   Table,
   Tag,
   Tbody,
@@ -32,8 +30,7 @@ import {
   Thead,
   Tr,
   UnorderedList,
-  chakra,
-  useToken
+  chakra
 } from '@chakra-ui/react';
 import {
   EmbeddableExplorer,
@@ -41,16 +38,16 @@ import {
   MultiCodeBlock,
   MultiCodeBlockContext
 } from '@apollo/chakra-helpers';
-import {FaGithub} from 'react-icons/fa';
-import {GatsbySeo} from 'gatsby-plugin-next-seo';
+import {FaDiscourse, FaGithub} from 'react-icons/fa';
+import {FiStar} from 'react-icons/fi';
+import {Link as GatsbyLink} from 'gatsby';
 import {Global} from '@emotion/react';
 import {MDXProvider} from '@mdx-js/react';
 import {MDXRenderer} from 'gatsby-plugin-mdx';
 import {PathContext, useFieldTableStyles} from '../utils';
-import {SIDEBAR_WIDTH_BASE, SIDEBAR_WIDTH_XL} from './Sidebar';
 import {TOTAL_HEADER_HEIGHT} from './Header';
 import {YouTube} from './YouTube';
-import {graphql, useStaticQuery} from 'gatsby';
+import {dirname, join} from 'path';
 import {kebabCase} from 'lodash';
 import {rehype} from 'rehype';
 import {useMermaidStyles} from '../utils/mermaid';
@@ -70,9 +67,6 @@ import 'prismjs/components/prism-swift';
 import 'prismjs/components/prism-tsx';
 import 'prismjs/components/prism-typescript';
 import 'prismjs/components/prism-yaml';
-import {ApolloHeader} from './ApolloHeader';
-import {ApolloSidebar} from './ApolloSidebar';
-import {ApolloTableOfContents} from './ApolloTableOfContents';
 
 const LIST_SPACING = 4;
 const HEADINGS = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'];
@@ -174,34 +168,10 @@ const {processSync} = rehype()
   });
 
 export default function Page({file, pageContext, uri}) {
-  const paddingTop = useToken('space', 10);
-  const paddingBottom = useToken('space', 12);
-  const scrollMarginTop = useMemo(
-    () => `calc(${paddingTop} + ${TOTAL_HEADER_HEIGHT}px)`,
-    [paddingTop]
-  );
-
   const [language, setLanguage] = useLocalStorage('language');
-  const [sidebarHidden, setSidebarHidden] = useLocalStorage('sidebar');
 
   const mermaidStyles = useMermaidStyles();
   const fieldTableStyles = useFieldTableStyles();
-
-  const {
-    site: {
-      siteMetadata: {siteUrl}
-    }
-  } = useStaticQuery(
-    graphql`
-      {
-        site {
-          siteMetadata {
-            siteUrl
-          }
-        }
-      }
-    `
-  );
 
   const {
     name,
@@ -214,9 +184,18 @@ export default function Page({file, pageContext, uri}) {
 
   const {frontmatter, headings} = childMdx || childMarkdownRemark;
   const {title, description, toc, tags} = frontmatter;
-  const {docset, versions, currentVersion, navItems, algoliaFilters} =
-    pageContext;
-  const titleFont = encodeURIComponent('Source Sans Pro');
+  const {versions, currentVersion, navItems} = pageContext;
+
+  const pageProps = usePageLayoutProps({
+    pageContext,
+    title,
+    description
+  });
+
+  const scrollMarginTop = useMemo(
+    () => `calc(${pageProps.paddingTop} + ${TOTAL_HEADER_HEIGHT}px)`,
+    [pageProps.paddingTop]
+  );
 
   const defaultVersion = useMemo(
     () => versions.find(version => !version.slug.includes('/')),
@@ -239,7 +218,7 @@ export default function Page({file, pageContext, uri}) {
     return (
       <Button
         as="a"
-        href={`${repo}/${path.join(...repoPath)}`}
+        href={`${repo}/${join(...repoPath)}`}
         variant="link"
         size="lg"
         leftIcon={<FaGithub />}
@@ -249,46 +228,8 @@ export default function Page({file, pageContext, uri}) {
     );
   }, [gitRemote, basePath, relativePath]);
 
-  const renderSwitcher = useCallback(
-    props => (
-      <DocsetMenu
-        docset={docset}
-        versions={versions}
-        currentVersion={currentVersion}
-        {...props}
-      />
-    ),
-    [docset, versions, currentVersion]
-  );
-
   return (
     <>
-      <GatsbySeo
-        title={title}
-        description={description}
-        canonical={siteUrl + uri}
-        openGraph={{
-          title,
-          description,
-          images: [
-            {
-              url: getShareImage({
-                title,
-                tagline: docset,
-                titleFont,
-                titleFontSize: 80,
-                titleExtraConfig: '_bold',
-                taglineFont: titleFont,
-                textColor: 'FFFFFF',
-                textLeftOffset: 80,
-                textAreaWidth: 1120,
-                cloudName: 'apollographql',
-                imagePublicID: 'apollo-docs-template2_dohzxt'
-              })
-            }
-          ]
-        }}
-      />
       <Global
         styles={{
           '.mermaid': {
@@ -304,48 +245,19 @@ export default function Page({file, pageContext, uri}) {
           path: name === 'index' ? uri : dirname(uri)
         }}
       >
-        <ApolloHeader
-          algoliaFilters={algoliaFilters}
-          navItems={navItems}
-          renderSwitcher={renderSwitcher({size: 'sm'})}
-          renderSwitcher1={renderSwitcher({d: {base: 'none', md: 'flex'}})}
-        />
-        <ApolloSidebar
-          in={sidebarHidden}
-          onClick={() => setSidebarHidden(false)}
-          navItems={navItems}
-          onHide={() => setSidebarHidden(true)}
-        />
-        <Box
-          marginLeft={{
-            base: 0,
-            md: sidebarHidden ? 0 : SIDEBAR_WIDTH_BASE,
-            xl: sidebarHidden ? 0 : SIDEBAR_WIDTH_XL
-          }}
-          transitionProperty="margin-left"
-          transitionDuration="normal"
-        >
-          {defaultVersion && defaultVersion.slug !== basePath && (
-            <VersionBanner
-              versionLabels={[defaultVersion.label, currentVersion]}
-              to={'/' + defaultVersion.slug}
-            />
-          )}
-          <Flex
-            maxW="6xl"
-            mx="auto"
-            align="flex-start"
-            px={{base: 6, md: 10}}
-            as="main"
-            sx={{
-              paddingTop,
-              paddingBottom
-            }}
-          >
-            <Box flexGrow="1" w="0">
-              <Heading as="h1" size="2xl">
-                {title}
-              </Heading>
+        <PageLayout
+          {...pageProps}
+          banner={
+            defaultVersion &&
+            defaultVersion.slug !== basePath && (
+              <VersionBanner
+                versionLabels={[defaultVersion.label, currentVersion]}
+                to={'/' + defaultVersion.slug}
+              />
+            )
+          }
+          subtitle={
+            <>
               {description && (
                 <chakra.h2
                   fontSize={{base: 'xl', md: '2xl'}}
@@ -357,98 +269,127 @@ export default function Page({file, pageContext, uri}) {
                 </chakra.h2>
               )}
               {tags?.length && (
-                <Flex gap={1} mt={{base: 2, md: 3}}>
-                  {tags.map(tag => (
-                    <PrimaryLink
-                      key={tag}
-                      href={`/technotes/tags/${kebabCase(tag)}`}
+                <HStack mt={{base: 2, md: 3}}>
+                  {tags.map((tag, index) => (
+                    <Tag
+                      key={index}
+                      as={GatsbyLink}
+                      to={`/technotes/tags/${kebabCase(tag)}`}
+                      size="lg"
                     >
-                      <Tag size="lg">{tag}</Tag>
-                    </PrimaryLink>
+                      {tag}
+                    </Tag>
                   ))}
-                </Flex>
+                </HStack>
               )}
-              <Divider my="8" />
-              <Box
-                fontSize={{md: 'lg'}}
-                lineHeight={{md: 1.7}}
-                css={{
-                  [HEADINGS]: {
-                    scrollMarginTop
-                  }
-                }}
-                sx={{
-                  [HEADINGS]: {
-                    a: {
-                      color: 'inherit'
-                    },
-                    code: {
-                      bg: 'none',
-                      p: 0,
-                      color: 'secondary'
-                    }
-                  },
-                  '>': {
-                    ':not(:last-child)': {
-                      mb: 6
-                    },
-                    [HEADINGS]: {
-                      ':not(:first-child)': {
-                        mt: 10,
-                        mb: 4
-                      }
-                    }
-                  },
-                  'img.screenshot': {
-                    shadow: 'md',
-                    rounded: 'md'
-                  },
-                  '.field-table': fieldTableStyles,
-                  '.sticky-table': {
-                    shadow: 'inner',
-                    table: {
-                      [['td', 'th']]: {
-                        ':first-of-type': {
-                          position: 'sticky',
-                          left: 0,
-                          bg: 'bg',
-                          borderRightWidth: 1
-                        }
-                      },
-                      'tr:last-child': {
-                        td: {
-                          borderBottom: 'none'
-                        }
-                      }
-                    }
-                  }
-                }}
+            </>
+          }
+          pagination={<Pagination navItems={navItems} />}
+          aside={
+            toc !== false && (
+              // hide the table of contents on the home page
+              <chakra.aside
+                d={{base: 'none', lg: 'flex'}}
+                flexDirection="column"
+                ml={{base: 10, xl: 16}}
+                w={250}
+                flexShrink="0"
+                pos="sticky"
+                top={scrollMarginTop}
+                maxH={`calc(100vh - ${scrollMarginTop} - ${pageProps.paddingBottom})`}
               >
-                <MultiCodeBlockContext.Provider value={{language, setLanguage}}>
-                  {childMdx ? (
-                    <MDXProvider components={mdxComponents}>
-                      <MDXRenderer>{childMdx.body}</MDXRenderer>
-                    </MDXProvider>
-                  ) : (
-                    processSync(childMarkdownRemark.html).result
-                  )}
-                </MultiCodeBlockContext.Provider>
-                <Box d={{lg: 'none'}}>{editOnGitHub}</Box>
-              </Box>
-              <Pagination navItems={navItems} />
-            </Box>
-            <ApolloTableOfContents
-              toc={toc}
-              top={scrollMarginTop}
-              paddingBottom={paddingBottom}
-              title={title}
-              headings={headings}
-              onClick={() => window.freddyWidget?.show()}
-              editOnGitHub={editOnGitHub}
-            />
-          </Flex>
-          <Footer />
-        </Box>
+                <Heading size="md" mb="3">
+                  {title}
+                </Heading>
+                <TableOfContents headings={headings} />
+                <Stack align="flex-start" spacing="3" mt="8">
+                  <Button
+                    onClick={() => window.freddyWidget?.show()}
+                    variant="link"
+                    size="lg"
+                    leftIcon={<FiStar />}
+                  >
+                    Rate article
+                  </Button>
+                  {editOnGitHub}
+                  <Button
+                    as="a"
+                    href="https://community.apollographql.com/"
+                    variant="link"
+                    size="lg"
+                    leftIcon={<FaDiscourse />}
+                  >
+                    Discuss in forums
+                  </Button>
+                </Stack>
+              </chakra.aside>
+            )
+          }
+          contentProps={{
+            css: {
+              [HEADINGS]: {
+                scrollMarginTop
+              }
+            },
+            sx: {
+              [HEADINGS]: {
+                a: {
+                  color: 'inherit'
+                },
+                code: {
+                  bg: 'none',
+                  p: 0,
+                  color: 'secondary'
+                }
+              },
+              '>': {
+                ':not(:last-child)': {
+                  mb: 6
+                },
+                [HEADINGS]: {
+                  ':not(:first-child)': {
+                    mt: 10,
+                    mb: 4
+                  }
+                }
+              },
+              'img.screenshot': {
+                shadow: 'md',
+                rounded: 'md'
+              },
+              '.field-table': fieldTableStyles,
+              '.sticky-table': {
+                shadow: 'inner',
+                table: {
+                  [['td', 'th']]: {
+                    ':first-of-type': {
+                      position: 'sticky',
+                      left: 0,
+                      bg: 'bg',
+                      borderRightWidth: 1
+                    }
+                  },
+                  'tr:last-child': {
+                    td: {
+                      borderBottom: 'none'
+                    }
+                  }
+                }
+              }
+            }
+          }}
+        >
+          <MultiCodeBlockContext.Provider value={{language, setLanguage}}>
+            {childMdx ? (
+              <MDXProvider components={mdxComponents}>
+                <MDXRenderer>{childMdx.body}</MDXRenderer>
+              </MDXProvider>
+            ) : (
+              processSync(childMarkdownRemark.html).result
+            )}
+          </MultiCodeBlockContext.Provider>
+          <Box d={{lg: 'none'}}>{editOnGitHub}</Box>
+        </PageLayout>
       </PathContext.Provider>
     </>
   );
